Split out into multiple classes
