+++ osu.Framework/Configuration/ISerializableBindable.cs
@@ -1,0 +1,17 @@
+﻿// Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
+// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
+
+using Newtonsoft.Json;
+
+namespace osu.Framework.Configuration
+{
+    /// <summary>
+    /// An interface which allows <see cref="Bindable{T}"/> to be json serialized/deserialized.
+    /// </summary>
+    [JsonConverter(typeof(BindableJsonConverter))]
+    internal interface ISerializableBindable
+    {
+        void SerializeTo(JsonWriter writer, JsonSerializer serializer);
+        void DeserializeFrom(JsonReader reader, JsonSerializer serializer);
+    }
+}
